requirements.txt
numpy~=1.20.3
opencv-python~=4.7.0.68
pandas~=1.4.2
imutils~=0.5.4
torch~=1.13.1
pillow~=8.4.0